Move tests to unit
