+++ Makefile
@@ -30,39 +30,42 @@ mkfs: $(SOURCE)/mkfs.c
 TEST=./test
 TEST_BIN=./test/bin
 
-test_disk_layer: test/test_disk_layer.c
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
+UNIT_TEST=./test/unit
+UNIT_TEST_BIN=./test/unit/bin
 
-test_superblock_layer: test/test_superblock_layer.c 
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
+test_disk_layer: $(UNIT_TEST)/test_disk_layer.c
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
 
-test_dblock_inode_layer: test/test_dblock_inode_layer.c
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
+test_superblock_layer: $(UNIT_TEST)/test_superblock_layer.c 
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
 
-test_dblock_inode_freelist: test/test_dblock_inode_freelist.c
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
+test_dblock_inode_layer: $(UNIT_TEST)/test_dblock_inode_layer.c
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
 
-test_inode_data_block_ops: test/test_inode_data_block_ops.c
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
+test_dblock_inode_freelist: $(UNIT_TEST)/test_dblock_inode_freelist.c
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
 
-test_directory_ops: test/test_directory_ops.c
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
+test_inode_data_block_ops: $(UNIT_TEST)/test_inode_data_block_ops.c
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
 
-test_interface_layer: test/test_interface_layer.c
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
+test_directory_ops: $(UNIT_TEST)/test_directory_ops.c
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
 
-test_interface_layer_disk: test/test_interface_layer.c
-	$(shell  mkdir -p $(TEST_BIN))
-	$(CC) -o $(TEST_BIN)/$@ $^ $(DEBUG_FLAGS) -DDISK_MEMORY
+test_interface_layer: $(UNIT_TEST)/test_interface_layer.c
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS)
 
-tests: test_disk_layer test_superblock_layer test_dblock_inode_layer test_dblock_inode_freelist test_inode_data_block_ops test_directory_ops test_interface_layer
+test_interface_layer_disk: $(UNIT_TEST)/test_interface_layer.c
+	$(shell  mkdir -p $(UNIT_TEST_BIN))
+	$(CC) -o $(UNIT_TEST_BIN)/$@ $^ $(DEBUG_FLAGS) -DDISK_MEMORY
+
+unit_tests: test_disk_layer test_superblock_layer test_dblock_inode_layer test_dblock_inode_freelist test_inode_data_block_ops test_directory_ops test_interface_layer
 
 # ============ CLEAN =============
