+++ test/unit/test_dblock_inode_freelist.c
@@ -3,9 +3,9 @@
 #include <errno.h>
 #include <stdlib.h>
 
-#include "../src/data_block_ops.c"
-#include "../src/disk_layer.c"
-#include "../src/superblock_layer.c"
+#include "../../src/data_block_ops.c"
+#include "../../src/disk_layer.c"
+#include "../../src/superblock_layer.c"
 
 #include "test_helpers.c"
