+++ test/unit/test_dblock_inode_layer.c
@@ -3,10 +3,10 @@
 #include <errno.h>
 #include <stdlib.h>
 
-#include "../src/disk_layer.c"
-#include "../src/superblock_layer.c"
-#include "../src/inode_ops.c"
-#include "../src/data_block_ops.c"
+#include "../../src/disk_layer.c"
+#include "../../src/superblock_layer.c"
+#include "../../src/inode_ops.c"
+#include "../../src/data_block_ops.c"
 
 #include "test_helpers.c"
