+++ test/unit/test_directory_ops.c
@@ -3,13 +3,13 @@
 #include <errno.h>
 #include <stdlib.h>
 
-#include "../src/disk_layer.c"
-#include "../src/superblock_layer.c"
-#include "../src/inode_ops.c"
-#include "../src/data_block_ops.c"
-#include "../src/inode_data_block_ops.c"
-#include "../src/inode_cache.c"
-#include "../src/directory_ops.c"
+#include "../../src/disk_layer.c"
+#include "../../src/superblock_layer.c"
+#include "../../src/inode_ops.c"
+#include "../../src/data_block_ops.c"
+#include "../../src/inode_data_block_ops.c"
+#include "../../src/inode_cache.c"
+#include "../../src/directory_ops.c"
 
 #include "test_helpers.c"
