+++ test/unit/test_inode_data_block_ops.c
@@ -3,11 +3,11 @@
 #include <errno.h>
 #include <stdlib.h>
 
-#include "../src/disk_layer.c"
-#include "../src/superblock_layer.c"
-#include "../src/inode_ops.c"
-#include "../src/data_block_ops.c"
-#include "../src/inode_data_block_ops.c"
+#include "../../src/disk_layer.c"
+#include "../../src/superblock_layer.c"
+#include "../../src/inode_ops.c"
+#include "../../src/data_block_ops.c"
+#include "../../src/inode_data_block_ops.c"
 
 #define INODE_DATA_BLOCK_OPS "altfs_inode_data_block_ops_test"
